==> 01 - Combined Dataset.ipynb <==
import pandas as pd
import requests
from bs4 import BeautifulSoup

from tqdm.notebook import tqdm

files = ['bproperty_bashundhara_RA.csv',
         'bproperty_dhanmondi.csv',
         'bproperty_gulshan.csv',
         'bproperty_mirpur.csv',
         'bproperty_uttara.csv' ]

dfs = []
for file in files:
    dfs.append(pd.read_csv(file))
df = pd.concat(dfs, ignore_index=True)

def get_unique_id(url):
    url_ = url
    unique = str(url_).replace('/en/property/details-', '').replace('.html', '')
    return unique

def fix_url(url):
    url = 'https://www.bproperty.com' + str(url)
    return url

df['unique_id'] = df['url'].apply(get_unique_id)
df['url'] = df['url'].apply(fix_url)

df.head()

df.shape

df['location'].value_counts()

secondary_scrape = {}

for i in tqdm(range(len(df))):
    url = df['url'][i]
    
    response = requests.get(url, headers={'User-Agent': 'Mozilla/5.0'})
    text = response.text
    soup = BeautifulSoup(text, 'lxml')
    
    features = soup.find_all("span", class_='_005a682a')
    
    apartment_features = {}
    
    for j in range(len(features)):
        features[j] = str(features[j]).replace('<span class="_005a682a">', '').replace('</span>', '').replace('<!-- -->', '').replace(': ', ':')
        try:
            key, value = features[j].split(':')
        except:
            key = features[j]
            value = 'Yes'
        apartment_features[key] = value
    features = list(set(features))
    
    secondary_scrape[ df['unique_id'][i] ] = apartment_features

scraped = pd.DataFrame(secondary_scrape)

scraped.shape

scraped = scraped.T

scraped.head()

scraped.to_csv('secondary.csv', index=False)

scraped.columns

scraped.fillna(0, inplace=True)

scraped

scraped = scraped.replace('Yes', 1)

scraped

scraped['Parking Spaces'].value_counts()

scraped[scraped['Parking Spaces'] == '90']

scraped['Elevators in Building'].value_counts()

scraped[scraped['Elevators in Building'] == '11']

scraped['Elevators in Building'] = scraped['Elevators in Building'].astype('int64')
scraped['Parking Spaces'] = scraped['Parking Spaces'].astype('int64')

scraped = scraped.reset_index()
scraped.rename(columns={'index':'unique_id'}, inplace=True)

scraped

scraped.info()

scraped.to_csv('secondary_formatted.csv', index=False)

def fix_dimensions(dim):
    dim = float(str(dim).replace(' sqft', '').replace(',', ''))
    return dim

df['dimensions'] = df['dimensions'].apply(fix_dimensions)

df['num_beds'] = df['num_beds'].astype('int64')
df['num_baths'] = df['num_baths'].astype('int64')

df

df.drop('description', axis=1, inplace=True)
df.drop('url', axis=1, inplace=True)

merged = pd.merge(df, scraped, on='unique_id')

merged.head(10)

merged.to_csv('dataset/merged.csv', index=False)
