==> 02 - Selenium Crawler for Important Local Places.ipynb <==
from selenium import webdriver 
import time
from tqdm.notebook import tqdm
import pandas as pd
import json, csv

import warnings
warnings.filterwarnings('ignore')

files = ['bproperty_bashundhara_RA.csv',
         'bproperty_dhanmondi.csv',
         'bproperty_gulshan.csv',
         'bproperty_mirpur.csv',
         'bproperty_uttara.csv' ]

dfs = []
for file in files:
    dfs.append(pd.read_csv(file))
df = pd.concat(dfs, ignore_index=True)

def get_unique_id(url):
    url_ = url
    unique = str(url_).replace('/en/property/details-', '').replace('.html', '')
    return unique

def fix_url(url):
    url = 'https://www.bproperty.com' + str(url)
    return url

df['unique_id'] = df['url'].apply(get_unique_id)
df['url'] = df['url'].apply(fix_url)

# df.to_csv('combined.csv', index=False)
df.head()

driver = webdriver.Edge(executable_path='driver/msedgedriver.exe')
tertiary_scrape = {}
bad_id = []

for i in tqdm(range(len(df))):
    try:
        url = df['url'][i]
        title = driver.get(url)

        # Click on JS Map
        driver.find_element_by_xpath('//div[@class="_4866b17e"]').click()
        time.sleep(3)

        # Click on required tabs and retrieve data
        schools = []
        restaurants = []
        hospitals = []
        parks = []

        tabs = driver.find_elements_by_xpath('//li[@class="a17d0d0d"]')

        for j in range(4):
            tabs[j].click()
            time.sleep(3)

            nearby = driver.find_elements(by='xpath', value='//div[@class="daf61ad1"]')

            if j == 0:
                for item in nearby:
                    schools.append(item.text)

            if j == 1:
                for item in nearby:
                    restaurants.append(item.text)

            if j == 2:
                for item in nearby:
                    hospitals.append(item.text)

            if j == 3:
                for item in nearby:
                    parks.append(item.text) 

        tertiary_scrape[ df['unique_id'][i] ] = {'schools': len(list(set(schools))),
                                                 'restaurants': len(list(set(restaurants))),
                                                 'hospitals': len(list(set(hospitals))),
                                                 'parks': len(list(set(parks)))}
        
        fields = [i, df['unique_id'][i], len(list(set(schools))), len(list(set(restaurants))),
                                      len(list(set(hospitals))), len(list(set(parks)))]
        
        with open(r'tertiary.csv', 'a') as f:
            writer = csv.writer(f)
            writer.writerow(fields)
            f.close()
            
        if i%100 == 0 and i!=0:
            print("Resting time: 10 mins (To avoid Bot Detection)")
            print()
            driver.quit()
            for s in tqdm(range(600)):
                time.sleep(1)
            driver = webdriver.Edge(executable_path='driver/msedgedriver.exe')
            
    except Exception as e:
        print('ID: ' + str(df['unique_id'][i]))
        bad_id.append(df['unique_id'][i])
        driver.quit()
        
        print("Retrying in: 10 mins")
        for s in tqdm(range(600)):
            time.sleep(1)
        
        
        print('Starting New Instance')
        print()
        
        driver = webdriver.Edge(executable_path='driver/msedgedriver.exe')
        

driver.quit()

tertiary_scrape

scraped = pd.DataFrame(tertiary_scrape)
scraped = scraped.T.reset_index()
scraped.rename(columns={'index':'unique_id'}, inplace=True)

scraped

bad_id
