==> 05 - Secondary Data Model.ipynb <==
import pandas as pd

df = pd.read_csv('dataset/merged.csv')

df.head()

df['location'].value_counts()

# df = df[ df['location'].isin(['Uttara', 'Bashundhara R-A', 'Dhanmondi']) ]

# Data Preprocessing

df.drop(['house_type', 'city', 'location', 'unique_id', 'area_block'], axis=1, inplace=True)

df.head()

# Feature Selection

df.shape

df.columns

df = df[ ['dimensions', 'num_beds', 'num_baths', 'price', 'View', 'Floor Level',
       'Balcony or Terrace', 'Flooring', 'Parking Spaces',
       'Elevators in Building', '24 Hours Concierge', 'Double Glazed Windows',
       'Freehold', 'Waste Disposal', 'Lobby in Building', 'Electricity Backup',
       'Intercom', 'CCTV Security', 'Maintenance Staff', 'Swimming Pool',
       'Storage Areas', 'Barbeque Area', 'Cleaning Services',
       'Service Elevators', 'Cafeteria or Canteen', 'Broadband Internet',
       'Central Heating', 'Lawn or Garden', 'Furnished', 'Prayer Room',
       'Conference Room', 'Business Center', 'Laundry Facility', 'Jacuzzi',
       'Steam Room', 'Day Care Center', 'First Aid Medical Center',
       'Shared Kitchen', 'ATM Facility', 'Facilities for Disabled', 'Sauna'] ].copy()
df = df.drop_duplicates()

df.shape

feature_col = ['dimensions', 'num_beds', 'num_baths', 'View', 'Floor Level',
       'Balcony or Terrace', 'Flooring', 'Parking Spaces',
       'Elevators in Building', '24 Hours Concierge', 'Double Glazed Windows',
       'Freehold', 'Waste Disposal', 'Lobby in Building', 'Electricity Backup',
       'Intercom', 'CCTV Security', 'Maintenance Staff', 'Swimming Pool',
       'Storage Areas', 'Barbeque Area', 'Cleaning Services',
       'Service Elevators', 'Cafeteria or Canteen', 'Broadband Internet',
       'Central Heating', 'Lawn or Garden', 'Furnished', 'Prayer Room',
       'Conference Room', 'Business Center', 'Laundry Facility', 'Jacuzzi',
       'Steam Room', 'Day Care Center', 'First Aid Medical Center',
       'Shared Kitchen', 'ATM Facility', 'Facilities for Disabled', 'Sauna']
target_col = ['price']

X = df[feature_col].copy()
y = df[target_col].copy()

# One Hot Encoding

# X = pd.get_dummies(X, prefix='area_block')

X.head()

# Feature Scaling

from sklearn.preprocessing import MinMaxScaler

scaler = MinMaxScaler()
X = scaler.fit_transform(X)

# XGBoost Regressor (Base)

from sklearn.model_selection import train_test_split

from xgboost import XGBRegressor

regressor = XGBRegressor()

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)

regressor.fit(X_train, y_train, early_stopping_rounds=10, eval_metric='mae', eval_set=[(X_test, y_test)])

y_pred = regressor.predict(X_test)

# Evaluation Metrics (Base)

from sklearn.metrics import mean_absolute_error, mean_squared_error, r2_score
import numpy as np 

print("Mean Absolute Error: ",mean_absolute_error(y_test,y_pred))
print("Mean Squared Error",mean_squared_error(y_test,y_pred))
print("Root MSE",np.sqrt(mean_squared_error(y_test,y_pred)))
print("RMS Log Error",np.log(np.sqrt(mean_squared_error(y_test,y_pred))))

r2 = r2_score(y_test,y_pred)
print("R Squared Error: ", r2)

scores = {}
scores['Base'] = {'Mean Absolute Error': mean_absolute_error(y_test,y_pred),
#                   'Mean Squared Error':mean_squared_error(y_test,y_pred),
                  'Root MSE':np.sqrt(mean_squared_error(y_test,y_pred)),
                  'RMS Log Error':np.log(np.sqrt(mean_squared_error(y_test,y_pred))),
                  'R Squared Error':r2_score(y_test,y_pred),
                 }

# Hyper Parameter Tuning

from sklearn.model_selection import RandomizedSearchCV

params = {
    'learning_rate': [0.05, 0.1, 0.15, 0.2, 0.25, 0.3],
    'max_depth': [3, 4, 5, 6, 8, 10, 12, 15],
    'min_child_weight': [1, 3, 5, 7],
    'gamma': [0, 0.1, 0.2, 0.3, 0.4],
    'colsample_bytree': [0.3, 0.4, 0.5, 0.7]
}

regressor = XGBRegressor(verbosity=0)

random_search = RandomizedSearchCV(regressor, param_distributions=params, n_iter=5, scoring='neg_mean_squared_error', n_jobs=-1, cv=5, verbose=3)

random_search.fit(X, y)

random_search.best_estimator_

regressor = XGBRegressor(base_score=0.5, booster='gbtree', colsample_bylevel=1,
             colsample_bynode=1, colsample_bytree=0.4, gamma=0.4, gpu_id=-1,
             importance_type='gain', interaction_constraints='',
             learning_rate=0.15, max_delta_step=0, max_depth=4,
             min_child_weight=5, monotone_constraints='()',
             n_estimators=100, n_jobs=12, num_parallel_tree=1, random_state=0,
             reg_alpha=0, reg_lambda=1, scale_pos_weight=1, subsample=1,
             tree_method='exact', validate_parameters=1, verbosity=0)

regressor.fit(X_train, y_train, early_stopping_rounds=10, eval_metric='mae', eval_set=[(X_test, y_test)])

y_pred = regressor.predict(X_test)

# Evaluation Metrics (Tuned)

from sklearn.metrics import mean_absolute_error, mean_squared_error, r2_score
import numpy as np 

print("Mean Absolute Error: ",mean_absolute_error(y_test,y_pred))
print("Mean Squared Error",mean_squared_error(y_test,y_pred))
print("Root MSE",np.sqrt(mean_squared_error(y_test,y_pred)))
print("RMS Log Error",np.log(np.sqrt(mean_squared_error(y_test,y_pred))))

r2 = r2_score(y_test,y_pred)
print("R Squared Error: ", r2)

scores['Tuned'] = {'Mean Absolute Error': round(mean_absolute_error(y_test,y_pred), 3),
#                   'Mean Squared Error':round(mean_squared_error(y_test,y_pred), 3),
                  'Root MSE':round(np.sqrt(mean_squared_error(y_test,y_pred)), 3),
                  'RMS Log Error':round(np.log(np.sqrt(mean_squared_error(y_test,y_pred))), 3),
                  'R Squared Error':round(r2_score(y_test,y_pred), 3),
                 }

eval_df = pd.DataFrame(scores)
eval_df
